Extract language detection partial
