+++ src/CodeIndex/Indexer/Scanning/FileIndexer.cs
@@ -30,215 +30,6 @@ public partial class FileIndexer
     private const int MaxProjectMarkerFingerprintDirectories = 8192;
     private const int MaxProjectMarkerFingerprintFiles = 4096;
     private const int MaxIgnorePatternLength = 512;
-    // Extension-to-language mapping / 拡張子→言語名マッピング
-    private static readonly Dictionary<string, string> LangMap = new(StringComparer.OrdinalIgnoreCase)
-    {
-        [".py"] = "python",
-        [".pyi"] = "python",  // Python type stub (PEP 561) / Python 型スタブ
-        [".pyw"] = "python",  // Windowed Python script / Windows 用 Python スクリプト
-        // Cython keeps its own bucket because it extends Python with native declarations
-        // (`cdef`, `cpdef`, `cimport`) that need Cython-specific symbol patterns.
-        // Cython は `cdef` / `cpdef` / `cimport` など Python を拡張した native 宣言を持つため、
-        // Python ではなく Cython 専用の symbol pattern を使う独立 bucket にする。
-        [".pyx"] = "cython",  // Cython source / Cython ソース
-        [".pxd"] = "cython",  // Cython declaration / Cython 宣言
-        [".js"] = "javascript",
-        [".cjs"] = "javascript",
-        [".mjs"] = "javascript",
-        [".ts"] = "typescript",
-        [".cts"] = "typescript",
-        [".mts"] = "typescript",
-        [".jsx"] = "javascript",
-        [".tsx"] = "typescript",
-        [".rb"] = "ruby",
-        [".rake"] = "ruby",    // Rake tasks / Rake タスク
-        [".gemspec"] = "ruby",    // RubyGems spec / RubyGems スペック
-        [".podspec"] = "ruby",    // CocoaPods spec (Ruby DSL) / CocoaPods スペック
-        [".groovy"] = "groovy",
-        [".gvy"] = "groovy",
-        [".gy"] = "groovy",
-        [".gsh"] = "groovy",
-        [".go"] = "go",
-        [".rs"] = "rust",
-        [".java"] = "java",
-        [".kt"] = "kotlin",
-        [".kts"] = "kotlin",  // Kotlin Script / Kotlin スクリプト (Gradle Kotlin DSL など)
-        [".swift"] = "swift",
-        [".cu"] = "cuda",
-        [".cuh"] = "cuda",
-        [".glsl"] = "glsl",
-        [".vert"] = "glsl",
-        [".frag"] = "glsl",
-        [".hlsl"] = "hlsl",
-        [".wgsl"] = "wgsl",
-        [".metal"] = "metal",
-        [".c"] = "c",
-        [".cpp"] = "cpp",
-        [".cc"] = "cpp",
-        [".cxx"] = "cpp",
-        [".h"] = "c",       // Could be C or C++; defaults to C for symbol extraction
-        [".hh"] = "cpp",
-        [".hpp"] = "cpp",
-        [".hxx"] = "cpp",
-        [".cs"] = "csharp",
-        [".cshtml"] = "csharp",  // Razor (ASP.NET MVC/Pages) / Razor テンプレート
-        [".razor"] = "csharp",  // Blazor component / Blazor コンポーネント
-        [".m"] = "objc",
-        [".mm"] = "objc",
-        [".php"] = "php",
-        [".s"] = "assembly", // Also used by Scheme; assembly is the more common default.
-        [".S"] = "assembly",
-        [".asm"] = "assembly",
-        [".nasm"] = "assembly",
-        [".sh"] = "shell",
-        [".sql"] = "sql",
-        [".pgsql"] = "sql",     // PostgreSQL dialect / PostgreSQL 方言
-        [".tsql"] = "sql",     // T-SQL (SQL Server) / T-SQL (SQL Server)
-        [".plsql"] = "sql",     // PL/SQL (Oracle) / PL/SQL (Oracle)
-        [".pls"] = "sql",     // PL/SQL script (Oracle) / PL/SQL スクリプト (Oracle)
-        [".pks"] = "sql",     // PL/SQL package spec (Oracle) / PL/SQL パッケージ仕様 (Oracle)
-        [".pkb"] = "sql",     // PL/SQL package body (Oracle) / PL/SQL パッケージ本体 (Oracle)
-        [".plb"] = "sql",     // PL/SQL wrapped source (Oracle) / PL/SQL ラップ済みソース (Oracle)
-        [".psql"] = "sql",     // psql scripts / psql スクリプト
-        [".md"] = "markdown",
-        [".yaml"] = "yaml",
-        [".yml"] = "yaml",
-        [".json"] = "json",
-        [".toml"] = "toml",
-        [".config"] = "xml",
-        [".runsettings"] = "xml",
-        [".rules"] = "config",
-        [".xaml"] = "xml",    // WPF/MAUI/Avalonia XAML / XAML テンプレート
-        [".axaml"] = "xml",    // Avalonia XAML / Avalonia XAML
-        [".sln"] = "solution", // Visual Studio solution / Visual Studio ソリューション
-        [".manifest"] = "app_manifest", // Windows application manifest / Windows アプリケーションマニフェスト
-        [".csproj"] = "msbuild",// C# project file / C# プロジェクトファイル
-        [".fsproj"] = "msbuild",// F# project file / F# プロジェクトファイル
-        [".vbproj"] = "msbuild",// VB.NET project file / VB.NET プロジェクトファイル
-        [".props"] = "msbuild",// MSBuild props / MSBuild プロパティ
-        [".targets"] = "msbuild",// MSBuild targets / MSBuild ターゲット
-        [".html"] = "html",
-        [".htm"] = "html",    // Legacy / Windows / IIS default / 旧来の Windows / IIS 既定拡張子
-        [".xhtml"] = "html",    // XHTML / XHTML
-        [".shtml"] = "html",    // Server-side includes / サーバサイドインクルード
-        [".css"] = "css",
-        [".scss"] = "css",
-        [".less"] = "css",    // Less preprocessor / Less プリプロセッサ
-        [".pcss"] = "css",    // PostCSS / PostCSS
-        // Sass indented syntax / Stylus use indentation instead of braces, so they live in
-        // separate buckets with conservative line-level extractors rather than the CSS
-        // brace-scoped extractor.
-        // Sass インデント構文と Stylus は波括弧ではなくインデントで構造化するため、
-        // CSS の波括弧スコープ抽出ではなく、保守的な行単位抽出を持つ別バケットに分ける。
-        [".sass"] = "sass",
-        [".styl"] = "stylus",
-        [".vue"] = "vue",
-        [".svelte"] = "svelte",
-        [".tf"] = "terraform",
-        [".v"] = "verilog",  // Verilog defaults here; SystemVerilog has its own extensions.
-        [".sv"] = "systemverilog",
-        [".svh"] = "systemverilog",
-        [".vhd"] = "vhdl",
-        [".vhdl"] = "vhdl",
-        [".lisp"] = "commonlisp",
-        [".lsp"] = "commonlisp",
-        [".cl"] = "commonlisp", // Common Lisp wins the default over OpenCL here.
-        [".rkt"] = "racket",
-        [".pas"] = "pascal",
-        [".pp"] = "pascal",
-        [".dpr"] = "pascal",
-        [".st"] = "smalltalk",
-        [".smalltalk"] = "smalltalk",
-        [".ada"] = "ada",
-        [".adb"] = "ada",
-        [".ads"] = "ada",
-        [".f"] = "fortran",
-        [".f77"] = "fortran",
-        [".f90"] = "fortran",
-        [".f95"] = "fortran",
-        [".f03"] = "fortran",
-        [".f08"] = "fortran",
-        [".for"] = "fortran",
-        [".ftn"] = "fortran",
-        [".cbl"] = "cobol",
-        [".cob"] = "cobol",
-        [".cobol"] = "cobol",
-        [".cpy"] = "cobol",   // COBOL copybook / COBOL コピー句
-        [".raku"] = "raku",
-        [".rakumod"] = "raku",
-        [".rakutest"] = "raku",
-        [".t"] = "perl",    // Common Perl test scripts / Perl の test スクリプト
-        [".dart"] = "dart",
-        [".scala"] = "scala",
-        [".sc"] = "scala",
-        [".r"] = "r",
-        [".R"] = "r",
-        [".ex"] = "elixir",
-        [".exs"] = "elixir",
-        [".lua"] = "lua",
-        [".ml"] = "ocaml",
-        [".mli"] = "ocaml",
-        [".cr"] = "crystal",
-        [".clj"] = "clojure",
-        [".cljs"] = "clojure",
-        [".cljc"] = "clojure",
-        [".edn"] = "clojure",
-        [".d"] = "d",
-        [".erl"] = "erlang",
-        [".hrl"] = "erlang",
-        [".jl"] = "julia",
-        [".nim"] = "nim",
-        [".nims"] = "nim",
-        [".pl"] = "perl",
-        [".pm"] = "perl",
-        [".pod"] = "perl",
-        [".psgi"] = "perl",
-        [".cgi"] = "perl",
-        [".fcgi"] = "perl",
-        [".t"] = "perl",
-        [".sol"] = "solidity",
-        [".tcl"] = "tcl",
-        [".tk"] = "tcl",
-        [".fs"] = "fsharp",
-        [".fsx"] = "fsharp",
-        [".fsi"] = "fsharp",
-        [".bas"] = "vb",
-        [".cls"] = "vb",
-        [".ctl"] = "vb",
-        [".dob"] = "vb",
-        [".dsr"] = "vb",
-        [".frm"] = "vb",
-        [".pag"] = "vb",
-        [".vba"] = "vb",
-        [".vb"] = "vb",
-        [".vbhtml"] = "vb",
-        [".vbs"] = "vb",
-        [".hs"] = "haskell",
-        [".lhs"] = "haskell",
-        [".zig"] = "zig",
-        [".proto"] = "protobuf",  // Protocol Buffers / Protocol Buffers 定義
-        [".graphql"] = "graphql",   // GraphQL schema/queries / GraphQL スキーマ・クエリ
-        [".gql"] = "graphql",
-        [".gradle"] = "gradle",    // Gradle build scripts / Gradle ビルドスクリプト
-        [".cmake"] = "cmake",     // CMake scripts / CMake スクリプト
-        [".mk"] = "makefile",  // Makefile fragment / Makefile フラグメント
-        [".ps1"] = "powershell",// PowerShell scripts / PowerShell スクリプト
-        [".psm1"] = "powershell",// PowerShell modules / PowerShell モジュール
-        [".psd1"] = "powershell",// PowerShell data files / PowerShell データファイル
-        [".bat"] = "batch",     // Windows batch files / Windows バッチファイル
-        [".cmd"] = "batch",
-        [".bash"] = "shell",
-        [".zsh"] = "shell",
-        [".fish"] = "shell",
-        [".dockerfile"] = "dockerfile", // Suffix-style Dockerfile names such as app.Dockerfile / app.Dockerfile 形式
-        [".containerfile"] = "dockerfile", // Suffix-style Containerfile names such as app.Containerfile / app.Containerfile 形式
-    };
-
-    private static readonly (string Pattern, string Language)[] DisplayOnlyLanguageExtensions =
-    [
-        (".S", "assembly"),
-    ];
-
     public const string MaxFileSizeEnvironmentVariable = "CDIDX_MAX_FILE_BYTES";
     // Default maximum file size to index (4 MiB). Larger generated/vendor payloads
     // can still be opted in with --max-file-bytes, but the default path should not
@@ -247,11 +38,6 @@ public partial class FileIndexer
     // --max-file-bytes で明示的に opt-in できるが、既定経路では一般的な source scan で
     // multi-MB の単一 byte[] を確保しない。
     public const long DefaultMaxFileSizeBytes = 4 * 1024 * 1024;
-    // Extensionless shebang detection reads at most the first physical line within this
-    // byte cap. NUL bytes or a line that reaches the cap without LF/CR are treated as
-    // unsupported so binary executables and minified data are not parsed as scripts.
-    private const int ShebangProbeByteLimit = 256;
-
     private readonly string _projectRoot;
     private readonly string _ignoreRuleRoot;
     private readonly IReadOnlyList<string> _ancestorIgnoreDirectories;
@@ -350,229 +136,6 @@ public partial class FileIndexer
         ExtractorPluginRegistry.LoadPatternConfigsForProjectRoot(_projectRoot);
         var pathComparer = _ignoreCase ? StringComparer.OrdinalIgnoreCase : StringComparer.Ordinal;
         (_submodulePaths, _submoduleAncestorPaths, _submoduleLoadWarnings) = LoadGitSubmodulePaths(_ignoreRuleRoot, _projectRoot, pathComparer);
-    }
-
-    /// <summary>
-    /// Return all file patterns (extensions and filenames) mapped to their language names.
-    /// 全ファイルパターン（拡張子とファイル名）と対応する言語名のマッピングを返す。
-    /// </summary>
-    public static IReadOnlyDictionary<string, string> GetLanguageExtensions()
-    {
-        // Merge extension map and filename map for a complete view
-        // 完全な一覧のため拡張子マップとファイル名マップを統合
-        var merged = new Dictionary<string, string>(StringComparer.Ordinal);
-        foreach (var (pattern, lang) in LangMap)
-            merged.TryAdd(pattern, lang);
-        // Keep display-only case variants that collapse in the case-insensitive detection map.
-        // case-insensitive な検出マップでは潰れる表示用 case variant を保持する。
-        foreach (var (pattern, lang) in DisplayOnlyLanguageExtensions)
-            merged.TryAdd(pattern, lang);
-        foreach (var (name, lang) in FileNameMap)
-            merged.TryAdd(name, lang);
-        // Surface suffixed variants like Dockerfile.dev / Makefile.am as `<Prefix><suffix>` entries
-        // so `cdidx languages` and the MCP listing reflect what TryDetectLanguage actually handles.
-        // Dockerfile.dev / Makefile.am のようなサフィックス付き変種も `<Prefix><suffix>` 形で
-        // 露出させ、`cdidx languages` や MCP の一覧が TryDetectLanguage の実挙動と一致するようにする。
-        foreach (var (prefix, lang) in FileNamePrefixMap)
-            merged.TryAdd($"{prefix}<suffix>", lang);
-        foreach (var (extension, lang) in ExtractorPluginRegistry.LanguageExtensions)
-            merged.TryAdd(extension, lang);
-        foreach (var (extension, lang) in LanguageMapOverrides.LoadEffectiveMap())
-            merged[extension] = lang;
-        return merged;
-    }
-
-    public static string? DetectLanguage(string filePath)
-        => TryDetectLanguage(filePath).Language;
-
-    internal static bool IsIgnoreFilePath(string path)
-        => IgnoreFileNames.Contains(Path.GetFileName(path), StringComparer.OrdinalIgnoreCase);
-
-    internal LanguageDetectionResult TryDetectLanguageForIndexing(
-        string filePath,
-        string? content = null,
-        FileProbeStatus? knownIndexability = null)
-        => TryDetectLanguage(filePath, content, _symlinkPolicy, _projectRoot, knownIndexability);
-
-    internal static string? GetReusableDetectedLanguage(
-        string filePath,
-        IReadOnlyDictionary<string, string>? detectedLanguages)
-    {
-        if (detectedLanguages == null || !detectedLanguages.TryGetValue(filePath, out var language))
-            return null;
-
-        return CanReuseDetectedLanguageWithoutContent(filePath, language)
-            ? language
-            : null;
-    }
-
-    internal static bool CanReuseDetectedLanguageWithoutContent(string filePath, string? language)
-    {
-        if (string.IsNullOrEmpty(language))
-            return false;
-
-        var fileName = Path.GetFileName(filePath);
-        if (FileNameMap.TryGetValue(fileName, out var nameLanguage))
-            return string.Equals(language, nameLanguage, StringComparison.Ordinal);
-
-        foreach (var (prefix, prefixLanguage) in FileNamePrefixMap)
-        {
-            if (fileName.Length > prefix.Length &&
-                fileName.StartsWith(prefix, StringComparison.OrdinalIgnoreCase))
-            {
-                return string.Equals(language, prefixLanguage, StringComparison.Ordinal);
-            }
-        }
-
-        if (TryDetectLanguageOverride(filePath, fileName, out var overrideLanguage))
-            return string.Equals(language, overrideLanguage, StringComparison.Ordinal);
-
-        var extension = Path.GetExtension(filePath);
-        return !string.IsNullOrEmpty(extension)
-            && !string.Equals(extension, ".h", StringComparison.OrdinalIgnoreCase);
-    }
-
-    internal static LanguageDetectionResult TryDetectLanguage(string filePath, string? content = null)
-        => TryDetectLanguage(filePath, content, SymlinkPolicy.None, projectRoot: null, knownIndexability: null);
-
-    internal static LanguageDetectionResult TryDetectLanguage(
-        string filePath,
-        string? content,
-        SymlinkPolicy symlinkPolicy,
-        string? projectRoot,
-        FileProbeStatus? knownIndexability = null)
-    {
-        // Exact filename matching beats extension lookup so manifest-style filenames like
-        // `pyproject.toml` can map to a dependency category instead of the generic file type.
-        // `pyproject.toml` のようなマニフェスト系ファイル名が、汎用拡張子ではなく
-        // dependency category に紐づくよう、完全一致ファイル名を拡張子より先に判定する。
-        var fileName = Path.GetFileName(filePath);
-        if (FileNameMap.TryGetValue(fileName, out var nameLang))
-            return new LanguageDetectionResult(FileProbeStatus.Supported, nameLang);
-
-        // Then try known filename prefixes for suffixed variants like Dockerfile.dev / Makefile.am.
-        // The suffix must be non-empty so a bare `Dockerfile.` with trailing dot does not match.
-        // Dockerfile.dev や Makefile.am のようなサフィックス付き変種を検出する。
-        // `Dockerfile.` のような末尾ドットだけの形には一致させないため、サフィックスは1文字以上必須。
-        foreach (var (prefix, prefixLang) in FileNamePrefixMap)
-        {
-            if (fileName.Length > prefix.Length &&
-                fileName.StartsWith(prefix, StringComparison.OrdinalIgnoreCase))
-            {
-                return new LanguageDetectionResult(FileProbeStatus.Supported, prefixLang);
-            }
-        }
-
-        var ext = Path.GetExtension(filePath);
-        if (TryDetectLanguageOverride(filePath, fileName, out var overrideLang))
-            return new LanguageDetectionResult(FileProbeStatus.Supported, overrideLang);
-
-        if (LangMap.TryGetValue(ext, out var lang))
-        {
-            if (lang == "c" && string.Equals(ext, ".h", StringComparison.OrdinalIgnoreCase) && !string.IsNullOrEmpty(content))
-            {
-                var cppHeaderLanguage = TryDetectCppHeaderLanguage(content);
-                if (cppHeaderLanguage != null)
-                    return new LanguageDetectionResult(FileProbeStatus.Supported, cppHeaderLanguage);
-            }
-
-            return new LanguageDetectionResult(FileProbeStatus.Supported, lang);
-        }
-
-        if (ExtractorPluginRegistry.LanguageExtensions.TryGetValue(ext, out var pluginLang))
-            return new LanguageDetectionResult(FileProbeStatus.Supported, pluginLang);
-
-        if (!string.IsNullOrEmpty(ext))
-        {
-            ExtractorPluginRegistry.LoadPatternConfigsForPath(filePath);
-            if (ExtractorPluginRegistry.LanguageExtensions.TryGetValue(ext, out pluginLang))
-                return new LanguageDetectionResult(FileProbeStatus.Supported, pluginLang);
-
-            return new LanguageDetectionResult(FileProbeStatus.Unsupported, null);
-        }
-
-        return TryDetectLanguageFromShebang(filePath, symlinkPolicy, projectRoot, knownIndexability);
-    }
-
-    private static bool TryDetectLanguageOverride(string filePath, string fileName, out string language)
-    {
-        language = string.Empty;
-        var overrides = LanguageMapOverrides.LoadEffectiveMap(filePath);
-        if (overrides.Count == 0)
-            return false;
-
-        for (var dotIndex = fileName.IndexOf('.'); dotIndex >= 0; dotIndex = fileName.IndexOf('.', dotIndex + 1))
-        {
-            if (overrides.TryGetValue(fileName[dotIndex..], out var mappedLanguage))
-            {
-                language = mappedLanguage;
-                return true;
-            }
-        }
-
-        return false;
-    }
-
-    private static string? TryDetectCppHeaderLanguage(string content)
-    {
-        const int maxLines = 200;
-        var remaining = content.AsSpan();
-        var inspectedLines = 0;
-
-        while (remaining.Length > 0 && inspectedLines < maxLines)
-        {
-            var newlineIndex = remaining.IndexOf('\n');
-            var line = newlineIndex >= 0 ? remaining[..newlineIndex] : remaining;
-            if (line.Length > 0 && line[^1] == '\r')
-                line = line[..^1];
-
-            if (LooksLikeCppHeaderLine(line))
-                return "cpp";
-
-            if (newlineIndex < 0)
-                break;
-
-            remaining = remaining[(newlineIndex + 1)..];
-            inspectedLines++;
-        }
-
-        return null;
-    }
-
-    private static bool LooksLikeCppHeaderLine(ReadOnlySpan<char> line)
-    {
-        line = line.Trim();
-        if (line.IsEmpty)
-            return false;
-
-        if (line.StartsWith("//") || line.StartsWith("/*") || line.StartsWith("*"))
-            return false;
-
-        if (line.StartsWith("namespace ", StringComparison.Ordinal)
-            || line.StartsWith("template ", StringComparison.Ordinal)
-            || line.StartsWith("template<", StringComparison.Ordinal)
-            || line.StartsWith("using ", StringComparison.Ordinal)
-            || line.StartsWith("class ", StringComparison.Ordinal)
-            || line.StartsWith("enum class ", StringComparison.Ordinal)
-            || line.StartsWith("enum struct ", StringComparison.Ordinal)
-            || line.StartsWith("public:", StringComparison.Ordinal)
-            || line.StartsWith("private:", StringComparison.Ordinal)
-            || line.StartsWith("protected:", StringComparison.Ordinal))
-        {
-            return true;
-        }
-
-        if (line.Contains("constexpr ", StringComparison.Ordinal)
-            || line.Contains("consteval ", StringComparison.Ordinal)
-            || line.Contains("constinit ", StringComparison.Ordinal)
-            || line.Contains("decltype(", StringComparison.Ordinal)
-            || line.Contains("friend ", StringComparison.Ordinal)
-            || line.Contains("std::", StringComparison.Ordinal))
-        {
-            return true;
-        }
-
-        return false;
     }
 
     internal static bool CanIndexFile(string filePath)
